tests: Move file hunk display coverage
The data package has a file hunk display module for file-scoped text
rendering, while hunk navigation tests still cover selected-state behavior.

The buffer-backed file view assertion belongs with the module that owns
file-scoped text hunk rendering.

This commit moves that assertion into test_file_hunk_display and removes
the hunk tracking copy.

Subsequent commits will move command imports onto the file hunk display
module.
